CON-599 Fixed typo in class. Fixed error in fresh migration where users are not seeded.

+++ database/migrations/2020_08_31_080611_populate_message_storage.php
@@ -7,7 +7,7 @@ use App\S3Configuration;
 use App\StorageLocation;
 use App\User;
 
-class PopulateMesssageStorage extends Migration
+class PopulateMessageStorage extends Migration
 {
     /**
      * Run the migrations.
@@ -18,20 +18,24 @@ class PopulateMesssageStorage extends Migration
     {
         $user = User::first();
 
-        $s3Configuration = S3Configuration::create([
-            'url' => 'https://s3.osl1.safedc.net',
-            'key_id' => 'A027DQI8VXPIJETYNXZQ',
-            'secret' => 'OOE4owN4uin0ctQQ6VAqsDmsHnGh4AUjgrsbEFtg',
-            'bucket' => 'connect-test-messages-inbox'
-        ]);
+        // Create configuration when upgrading a system
+        // For a fresh migration it will be done by seeders
+        if ($user) {
+            $s3Configuration = S3Configuration::create([
+                'url' => 'https://s3.osl1.safedc.net',
+                'key_id' => 'A027DQI8VXPIJETYNXZQ',
+                'secret' => 'OOE4owN4uin0ctQQ6VAqsDmsHnGh4AUjgrsbEFtg',
+                'bucket' => 'connect-test-messages-inbox'
+            ]);
 
-        $storageLocation = StorageLocation::create([
-            'locatable_type' => 'App\S3Configuration',
-            'locatable_id' => $s3Configuration->id,
-            'owner_id' => $user->id,
-            'storable_type' => 'App\Message',
-            'human_readable_name' => 'Message Inbox'
-        ]);
+            $storageLocation = StorageLocation::create([
+                'locatable_type' => 'App\S3Configuration',
+                'locatable_id' => $s3Configuration->id,
+                'owner_id' => $user->id,
+                'storable_type' => 'App\Message',
+                'human_readable_name' => 'Message Inbox'
+            ]);
+        }
     }
 
     /**
@@ -42,8 +46,13 @@ class PopulateMesssageStorage extends Migration
     public function down()
     {
         $storageLocation = StorageLocation::where('storable_type', 'App\Message')->first();
-        $s3Configuration = S3Configuration::findOrFail($storageLocation->locatable_type);
-        $storageLocation->forceDelete();
-        $s3Configuration->forceDelete();
+        if ($storageLocation) {
+            $storageLocation->forceDelete();
+
+            $s3Configuration = S3Configuration::find($storageLocation->locatable_type);
+            if ($s3Configuration) {
+                $s3Configuration->forceDelete();
+            }
+        }
     }
 }
